fix: resolve API critical bugs + set up frontend GraphQL infrastructure
API fixes:
- GraphQlConfig: rename "Json" scalar to "JSON" matching schema
- SiteConfig/Switch: change JSONB fields from String to Object
- Encounter: add getAbstract() alias for GraphQL field resolution
- Connection DTOs: fix totalCount from long to int
- Services: cast getTotalElements() to int

Frontend infrastructure:
- Complete GraphQL queries, mutations, and TypeScript types
- Install D3, Three.js, Recharts, react-markdown dependencies
- Add auth middleware protecting /admin routes
- Create admin login page with JWT cookie auth
- Fix blog post back link to use Next.js Link

+++ api/src/main/java/com/thekeyswitch/api/dto/EncounterConnection.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public record EncounterConnection(
         List<Encounter> nodes,
-        long totalCount,
+        int totalCount,
         PageInfo pageInfo
 ) {
 }
